+++ packages/data-schema/src/MappedTypes/ResolveFieldProperties.ts
@@ -168,11 +168,13 @@ export type ResolveRefsOfCustomType<
     ? ResolveRef<NonModelTypes, R>
     : T[Prop];
 } extends infer Resolved
-  ? Intersection<
-      ExtractNullableFieldsToOptionalFields<Resolved>,
-      ExtractNonNullableFieldsToRequiredFields<Resolved>
-    >
+  ? ResolveCustomTypeFieldsRequirements<Resolved>
   : never;
+
+export type ResolveCustomTypeFieldsRequirements<Resolved> = Intersection<
+  ExtractNullableFieldsToOptionalFields<Resolved>,
+  ExtractNonNullableFieldsToRequiredFields<Resolved>
+>;
 
 type ResolveRelationships<
   Schema,